refactor: remove DatabaseTrieWitness trait, inline overlay_witness logic
Remove the DatabaseTrieWitness trait and its implementation from
reth-trie-db. The overlay_witness and from_tx logic is now inlined
directly at each call site in the state providers and tests.

Amp-Thread-ID: https://ampcode.com/threads/T-019c9422-97c3-70ed-802a-0f42d232f2de
