+++ crates/trie/db/src/lib.rs
@@ -10,7 +10,6 @@ mod proof;
 mod state;
 mod storage;
 mod trie_cursor;
-mod witness;
 
 pub use hashed_cursor::{
     DatabaseHashedAccountCursor, DatabaseHashedCursorFactory, DatabaseHashedStorageCursor,
@@ -24,7 +23,6 @@ pub use trie_cursor::{
     DatabaseAccountTrieCursor, DatabaseStorageTrieCursor, DatabaseTrieCursorFactory,
     LegacyKeyAdapter, PackedKeyAdapter, StorageTrieEntryLike, TrieKeyAdapter, TrieTableAdapter,
 };
-pub use witness::DatabaseTrieWitness;
 
 /// Dispatches a trie operation using the correct [`TrieKeyAdapter`] based on storage settings.
 ///
